requirements.txt
fastapi==0.104.1
uvicorn==0.23.2
gunicorn==21.2.0
httpx==0.25.1
boto3==1.29.0
pydantic==2.4.2
python-multipart==0.0.6
python-jose==3.3.0
cryptography==41.0.5
python-dotenv==1.0.0
email-validator==2.1.0